refactor: rename wallet-on-chain-address domain
